+++ Realm/ObjectServerTests/SwiftUIServerTests.swift
@@ -287,9 +287,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
                                 configuration: configuration,
                                 timeout: timeout)
         autoOpen.projectedValue
-            .sink { autoOpenState in
-                handler(autoOpenState)
-            }
+            .sink(receiveValue: handler)
             .store(in: &cancellables)
         waitForExpectations(timeout: 10.0)
         autoOpen.cancel()
@@ -365,6 +363,53 @@ class SwiftUIServerTests: SwiftSyncTestCase {
                 ex.fulfill()
             }
         }
+
+        proxy.stop()
+    }
+
+    // In case of no internet connection AutoOpen should return an opened Realm, offline-first approach
+    func testAutoOpenOpenForFlexibleSyncConfigWithoutInternetConnection() throws {
+        let proxy = TimeoutProxyServer(port: 5678, targetPort: 9090)
+        try proxy.start()
+
+        let appConfig = AppConfiguration(baseURL: "http://localhost:5678",
+                                         transport: AsyncOpenConnectionTimeoutTransport(),
+                                         localAppName: nil,
+                                         localAppVersion: nil)
+
+        try autoreleasepool {
+            try populateFlexibleSyncData { realm in
+                for i in 1...10 {
+                    // Using firstname to query only objects from this test
+                    let person = SwiftPerson(firstName: "\(#function)",
+                                             lastName: "lastname_\(i)",
+                                             age: i)
+                    realm.add(person)
+                }
+            }
+        }
+        App.resetAppCache()
+
+        let app = App(id: flexibleSyncAppId, configuration: appConfig)
+
+        let user = try logInUser(for: basicCredentials(app: app), app: app)
+        var configuration = user.flexibleSyncConfiguration()
+        configuration.objectTypes = [SwiftPerson.self]
+
+        proxy.dropConnections = true
+        let ex = expectation(description: "download-realm-flexible-auto-open-no-connection")
+        let autoOpen = AutoOpen(appId: flexibleSyncAppId, configuration: configuration, timeout: 1000)
+        autoOpen.projectedValue
+            .sink { autoOpenState in
+                if case let .open(realm) = autoOpenState {
+                    XCTAssertTrue(realm.isEmpty) // should not have downloaded anything
+                    ex.fulfill()
+                }
+            }
+            .store(in: &cancellables)
+
+        waitForExpectations(timeout: 10.0)
+        autoOpen.cancel()
 
         proxy.stop()
     }